requirements.txt
flask
requests
openai
python-telegram-bot
schedule
